+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - IStorageHandler signature to use initialize and upgrade methods
 
+### Fixed
+- Export IStorageHandler interface
+
 ## [1.0.0-rc.1] - 2022-10-12
 ### Added
 - Initial release